chore: refactor independent invoice and derive crates
