+++ descriptors/Cargo.toml
@@ -17,11 +17,11 @@ name = "descriptors"
 
 [dependencies]
 amplify = { workspace = true }
-bp-std = { workspace = true }
+bp-derive = { workspace = true }
 indexmap = { workspace = true }
 serde_crate = { workspace = true, optional = true }
 
 [features]
 default = []
 all = ["serde"]
-serde = ["serde_crate", "bp-std/serde"]
+serde = ["serde_crate", "bp-derive/serde"]
