+++ invoice/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "bp-invoice"
+description = "Bitcoin addresses and invoices"
+readme = "../README.md"
+version.workspace = true
+keywords.workspace = true
+categories.workspace = true
+authors.workspace = true
+homepage.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[lib]
+name = "invoice"
+
+[dependencies]
+amplify = { workspace = true }
+bech32 = { workspace = true }
+bitcoin_hashes = { workspace = true }
+bp-consensus = { workspace = true }
+serde_crate = { workspace = true, optional = true }
+
+[features]
+default = []
+all = []
+serde = ["serde_crate", "bp-consensus/serde"]
