+++ invoice/src/lib.rs
@@ -20,29 +20,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use derive::{Address, AddressError, AddressNetwork, DeriveScripts, Idx, Keychain, NormalIndex};
+#[macro_use]
+extern crate amplify;
+#[cfg(feature = "serde")]
+extern crate serde_crate as serde;
 
-#[derive(Clone, Eq, PartialEq, Hash, Debug)]
-pub struct AddressFactory<D: DeriveScripts> {
-    pub descriptor: D,
-    pub network: AddressNetwork,
-    pub keychain: Keychain,
-    pub unused_tip: NormalIndex,
-}
+pub mod base58;
+mod address;
+mod network;
 
-impl<D: DeriveScripts> Iterator for AddressFactory<D> {
-    type Item = Address;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let addr =
-            self.descriptor.derive_address(self.network, self.keychain, self.unused_tip).ok()?;
-        self.unused_tip.wrapping_inc_assign();
-        Some(addr)
-    }
-}
-
-impl<D: DeriveScripts> AddressFactory<D> {
-    pub fn address(&self, index: NormalIndex) -> Result<Address, AddressError> {
-        self.descriptor.derive_address(self.network, self.keychain, index)
-    }
-}
+pub use address::{
+    Address, AddressError, AddressNetwork, AddressParseError, AddressPayload, AddressType,
+};
+pub use network::{Network, UnknownNetwork};
