+++ invoice/src/network.rs
