+++ psbt/Cargo.toml
@@ -14,7 +14,7 @@ license = { workspace = true }
 
 [dependencies]
 amplify = { workspace = true }
-bp-std = { workspace = true }
+bp-derive = { workspace = true }
 descriptors = { workspace = true }
 indexmap = { workspace = true }
 base64 = "0.21.5"
@@ -24,4 +24,4 @@ serde_crate = { workspace = true, optional = true }
 [features]
 default = []
 all = ["serde"]
-serde = ["serde_crate", "bp-std/serde", "indexmap/serde"]
+serde = ["serde_crate", "bp-derive/serde", "indexmap/serde"]
