--- std/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "bp-std"
-version = { workspace = true }
-description = "Modern & lightweight implementation of bitcoin standards without rust-bitcoin/miniscript dependencies"
-keywords = { workspace = true }
-categories = { workspace = true }
-readme = "../README.md"
-authors = { workspace = true }
-homepage = { workspace = true }
-repository = { workspace = true }
-rust-version = { workspace = true }
-edition = { workspace = true }
-license = { workspace = true }
-
-[lib]
-name = "bpstd"
-
-[dependencies]
-amplify = { workspace = true }
-bitcoin_hashes = { workspace = true }
-bech32 = { workspace = true }
-commit_verify = "0.11.0-beta.1"
-bp-consensus = { workspace = true }
-bp-core = { version = "0.11.0-beta.1", optional = true }
-indexmap = { workspace = true }
-serde_crate = { workspace = true, optional = true }
-
-[features]
-default = []
-all = ["core", "serde"]
-core = ["bp-core"]
-serde = ["serde_crate", "bp-consensus/serde", "bp-core/serde"]
